+++ Api/src/main/java/com/Servify/controller/DTOServify.java
@@ -1,0 +1,5 @@
+package com.Servify.controller;
+
+public interface DTOServify {
+    void assertEmpty() throws EmptyDTOError;
+}
