+++ Api/src/main/java/com/Servify/model/CategoryService.java
@@ -12,7 +12,6 @@ public class CategoryService {
     @Id
     @Column
     private String category;
-
     public CategoryService() {
     }
